backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
weasyprint==62.1
jinja2==3.1.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
reportlab==4.2.0
